requirements.txt
einops==0.4.0
load-confounds==0.12.0
nibabel==3.2.2
nilearn==0.9.0
matplotlib==3.5.1
numpy==1.19.5
pandas==1.1.5
Pillow==9.0.1
pytest==7.0.1
seaborn==0.11.2
templateflow==0.7.2
# torch==1.11
torchviz==0.0.2
transformers==4.16.2
tqdm==4.62.3
wandb==0.12.10
webdataset==0.1.103